% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalReportRS.R
\name{survivalReportRS}
\alias{survivalReportRS}
\title{survivalReportRS creates a survival report}
\usage{
survivalReportRS()
}
\value{
the report
}
\description{
this function creates a survival report
}
\author{
Wheater, S.
}
